Fix read and write return types: now st_size_t, like the size inputs.
To support this change, turn on lots more compiler warnings, and fix
all the warnings generated.

+++ src/g72x.c
@@ -373,7 +373,7 @@ void update(int code_size, int y, int wi, int fi, int dq, int sr,
                 state_ptr->dq[cnt] = state_ptr->dq[cnt-1];
         /* FLOAT A : convert dq[0] to 4-bit exp, 6-bit mantissa f.p. */
         if (mag == 0) {
-                state_ptr->dq[0] = (dq >= 0) ? 0x20 : 0xFC20;
+                state_ptr->dq[0] = (dq >= 0) ? 0x20 : (short)0xFC20;
         } else {
                 exp = quan(mag, power2, 15);
                 state_ptr->dq[0] = (dq >= 0) ?
@@ -393,7 +393,7 @@ void update(int code_size, int y, int wi, int fi, int dq, int sr,
                 exp = quan(mag, power2, 15);
                 state_ptr->sr[0] =  (exp << 6) + ((mag << 6) >> exp) - 0x400;
         } else
-                state_ptr->sr[0] = 0xFC20;
+                state_ptr->sr[0] = (short)(0xFC20);
 
         /* DELAY A */
         state_ptr->pk[1] = state_ptr->pk[0];
